+++ packages/rainbowkit/src/transactions/transactionStore.ts
@@ -1,4 +1,4 @@
-import type { Address, PublicClient } from 'viem';
+import type { Address, PublicClient, TransactionReceipt } from 'viem';
 
 const storageKey = 'rk-transactions';
 
@@ -67,6 +67,9 @@ export function createTransactionStore({
 
   let provider = initialProvider;
   const listeners: Set<() => void> = new Set();
+  const transactionListeners: Set<
+    (txStatus: TransactionReceipt['status']) => void
+  > = new Set();
   const transactionRequestCache: Map<string, Promise<void>> = new Map();
 
   function setProvider(newProvider: PublicClient): void {
@@ -153,6 +156,8 @@ export function createTransactionStore({
                 // @ts-ignore - types changed with viem@1.1.0
                 status === 0 || status === 'reverted' ? 'failed' : 'confirmed',
               );
+
+              notifyTransactionListeners(status);
             })
             .catch(() => {
               // If a transaction is not found or cancelled
@@ -207,6 +212,14 @@ export function createTransactionStore({
     }
   }
 
+  function notifyTransactionListeners(
+    txStatus: TransactionReceipt['status'],
+  ): void {
+    for (const transactionListener of transactionListeners) {
+      transactionListener(txStatus);
+    }
+  }
+
   function onChange(fn: () => void): () => void {
     listeners.add(fn);
 
@@ -215,10 +228,21 @@ export function createTransactionStore({
     };
   }
 
+  function onTransactionStatus(
+    fn: (txStatus: TransactionReceipt['status']) => void,
+  ): () => void {
+    transactionListeners.add(fn);
+
+    return () => {
+      transactionListeners.delete(fn);
+    };
+  }
+
   return {
     addTransaction,
     clearTransactions,
     getTransactions,
+    onTransactionStatus,
     onChange,
     setProvider,
     waitForPendingTransactions,