+++ examples/with-create-react-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # with-create-react-app
 
+## 0.1.86
+
+### Patch Changes
+
+- Updated dependencies [af4ea4e]
+- Updated dependencies [f0b3b25]
+  - @rainbow-me/rainbowkit@2.0.7
+
 ## 0.1.85
 
 ### Patch Changes