+++ examples/with-next-rainbow-button/CHANGELOG.md
@@ -1,5 +1,11 @@
 # with-next-rainbow-button
 
+## 0.0.16
+
+### Patch Changes
+
+- @rainbow-me/rainbow-button@0.2.7
+
 ## 0.0.15
 
 ### Patch Changes