+++ examples/with-vite/CHANGELOG.md
@@ -1,5 +1,13 @@
 # with-vite
 
+## 0.0.75
+
+### Patch Changes
+
+- Updated dependencies [af4ea4e]
+- Updated dependencies [f0b3b25]
+  - @rainbow-me/rainbowkit@2.0.7
+
 ## 0.0.74
 
 ### Patch Changes