+++ packages/rainbow-button/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rainbow-me/rainbow-button
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [af4ea4e]
+- Updated dependencies [f0b3b25]
+  - @rainbow-me/rainbowkit@2.0.7
+
 ## 0.2.6
 
 ### Patch Changes